+++ ceammc/ext/src/base/mod_base.cpp
@@ -52,6 +52,7 @@ void setup_prop_get();
 void setup_prop_get_tilde();
 void setup_prop_join();
 void setup_prop_set();
+void setup_prop_split();
 void setup_snd_file();
 
 void ceammc_base_setup()
@@ -104,6 +105,7 @@ void ceammc_base_setup()
     setup_prop_get_tilde();
     setup_prop_join();
     setup_prop_set();
+    setup_prop_split();
     setup_snd_file();
 
     setup_test0x2edata();