+++ ceammc/CHANGELOG.md
@@ -26,6 +26,7 @@
     - prop.set renamed to prop.join (with prop<- alias)
     - prop.set now is a new object with different behavior, it changes properties via internal pd calls
     - prop.set alias added: p.set
+    - prop.get renamed to prop.split (with prop<- and @<- aliases)
 - spat.pan4~ and spat.pan8~:
     - @dist property renamed to @radius
     - positional args order reversed, now is: radius, angle